refactor: map features
move map-context from ui to features and reorganize de files
