+++ apps/search/src/app/app.module.ts
@@ -2,6 +2,7 @@ import { HttpClientModule } from '@angular/common/http'
 import { NgModule } from '@angular/core'
 import { BrowserModule } from '@angular/platform-browser'
 import { FeatureCatalogModule } from '@geonetwork-ui/feature/catalog'
+import { FeatureMapModule } from '@geonetwork-ui/feature/map'
 import { UiLayoutModule } from '@geonetwork-ui/ui/layout'
 import { UiMapModule } from '@geonetwork-ui/ui/map'
 import {
@@ -37,7 +38,7 @@ export const metaReducers: MetaReducer<any>[] = !environment.production
     FeatureSearchModule,
     FeatureCatalogModule,
     UiLayoutModule,
-    UiMapModule,
+    FeatureMapModule,
     StoreModule.forRoot({}, { metaReducers }),
     !environment.production ? StoreDevtoolsModule.instrument() : [],
     EffectsModule.forRoot(),
