+++ libs/feature/map/.eslintrc.json
@@ -13,7 +13,7 @@
           "error",
           {
             "type": "attribute",
-            "prefix": "geonetworkUi",
+            "prefix": "gnUi",
             "style": "camelCase"
           }
         ],
@@ -21,7 +21,7 @@
           "error",
           {
             "type": "element",
-            "prefix": "geonetwork-ui",
+            "prefix": "gn-ui",
             "style": "kebab-case"
           }
         ]
