+++ libs/feature/map/src/index.ts
@@ -1,1 +1,4 @@
 export * from './lib/feature-map.module'
+export * from './lib/manager/map-manager.service'
+export * from './lib/feature-info/feature-info.service'
+export * from './lib/map-context/map-context.model'
