+++ libs/ui/map/src/index.ts
@@ -1,4 +1,1 @@
 export * from './lib/ui-map.module'
-export * from './lib/models'
-export * from './lib/services'
-export * from './lib/fixtures'
