--- libs/ui/map/src/lib/fixtures/index.ts
@@ -1,2 +1,0 @@
-export * from './geojson.fixtures'
-export * from './map-context.fixtures'
