--- libs/ui/map/src/lib/models/index.ts
@@ -1,1 +1,0 @@
-export * from './map-context.model'
