--- libs/ui/map/src/lib/services/index.ts
@@ -1,1 +1,0 @@
-export * from './map-context.service'
