+++ libs/ui/map/src/lib/ui-map.module.ts
@@ -2,11 +2,10 @@ import { HttpClientModule } from '@angular/common/http'
 import { NgModule } from '@angular/core'
 import { CommonModule } from '@angular/common'
 import { MapComponent } from './components/map/map.component'
-import { MapContextComponent } from './components/map-context/map-context.component'
 
 @NgModule({
-  declarations: [MapComponent, MapContextComponent],
+  declarations: [MapComponent],
   imports: [CommonModule, HttpClientModule],
-  exports: [MapComponent, MapContextComponent],
+  exports: [MapComponent],
 })
 export class UiMapModule {}
