+++ libs/util/shared/src/index.ts
@@ -2,3 +2,4 @@ export * from './lib/util-shared.module'
 export * from './lib/services'
 export * from './lib/models/'
 export * from './lib/utils/'
+export * from './lib/fixtures/'
